+++ src/components/Question.vue
@@ -5,7 +5,6 @@
     </div>
     <div class="col-9">
       <router-view></router-view>
-      <button @click="goCreateQuestion" class="btn btn-info btn-lg fixed-bottom ml-auto m-5">質問する</button>
     </div>
   </div>
 </template>
@@ -13,8 +12,8 @@
 <script>
 import { reactive } from "vue";
 // import axios from "axios";
-import { useRouter } from "vue-router";
-import { useStore } from "vuex";
+// import { useRouter } from "vue-router";
+// import { useStore } from "vuex";
 
 import TagList from "./TagList.vue";
 
@@ -25,23 +24,14 @@ export default {
   },
   props: {},
   setup() {
-    const router = useRouter();
+    // const router = useRouter();
     // const route = useRoute();
-    const store = useStore();
+    // const store = useStore();
 
     const data = reactive({});
 
-    function goCreateQuestion() {
-      store.commit("resetAlert");
-      store.commit("resetQuestionDetails");
-      router.push({
-        name: "post",
-      });
-    }
-
     return {
       data,
-      goCreateQuestion,
     };
   },
 };
